Initial 9 LED fade program test
Experimented with the PWM functions on the various pins of the ATtiny202 although not all pins are PWM enabled. Potential possibility to enable all 12 LED's with PWM function in the future by switching pins being used for respective operations allowing all I/O pins handling the charlieplexed configuration to be PWM pins.